chore(assert): move `_format.ts`
