+++ assert/assert_equals.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 import { equal } from "./equal.ts";
-import { format } from "./_format.ts";
+import { format } from "../_util/_format.ts";
 import { AssertionError } from "./assertion_error.ts";
 import { red } from "../fmt/colors.ts";
 import { buildMessage, diff, diffstr } from "./_diff.ts";
