+++ assert/assert_greater_or_equal.ts
@@ -1,5 +1,5 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
-import { format } from "./_format.ts";
+import { format } from "../_util/_format.ts";
 import { AssertionError } from "./assertion_error.ts";
 
 /**
