+++ assert/assert_not_strict_equals.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 import { AssertionError } from "./assertion_error.ts";
-import { format } from "./_format.ts";
+import { format } from "../_util/_format.ts";
 
 /**
  * Make an assertion that `actual` and `expected` are not strictly equal.
